+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v1.0.9-alpha.3.240305
 ### 2024/3/5
 - issue bug修复
+- chrome driver path 自定义
+- os selenium spider 环境检测
 
 ## v1.0.9-alpha.2.240304
 ### 2024/3/4